requirements.txt
Django==1.9
Pillow==2.3.0
beautifulsoup4==4.6.0
chardet==2.2.1
colorama==0.2.5
decorator==3.4.0
django-filter==1.1.0
djangorestframework==3.7.3
html5lib==0.999
ipython==1.2.1
matplotlib==1.3.1
nose==1.3.1
numpy==1.8.2
pycurl==7.19.3
pygobject==3.12.0
pyparsing==2.0.1
python-apt==0.9.3.5ubuntu2
python-dateutil==2.0
pytz==2012c
requests==2.2.1
scipy==0.13.3
simplegeneric==0.8.1
six==1.5.2
tornado==3.1.1
unattended-upgrades==0.1
urllib3==1.7.1
wheel==0.24.0
